feat: adding project section
