+++ server/requirements.txt
@@ -13,4 +13,4 @@ beautifulsoup4==4.12.3
 langchain==0.2.1
 langchain-community==0.2.1
 langchain-google-genai==1.0.5
-pymongo
+pymongo==4.10.1
